pebble: extract bench code
Move the `pebble bench` code to a separate `./bench` package.
