+++ bench/sync.go
@@ -1,0 +1,132 @@
+// Copyright 2018 The LevelDB-Go and Pebble Authors. All rights reserved. Use
+// of this source code is governed by a BSD-style license that can be found in
+// the LICENSE file.
+
+package bench
+
+import (
+	"fmt"
+	"log"
+	"math/rand/v2"
+	"sync"
+	"sync/atomic"
+	"time"
+
+	"github.com/cockroachdb/pebble"
+	"github.com/cockroachdb/pebble/cockroachkvs"
+	"github.com/cockroachdb/pebble/internal/randvar"
+)
+
+// SyncConfig configures the sync benchmark.
+type SyncConfig struct {
+	Batch   *randvar.Flag
+	WALOnly bool
+	Values  *randvar.BytesFlag
+}
+
+// DefaultSyncConfig returns a SyncConfig populated with the same defaults as
+// the cmd/pebble CLI.
+func DefaultSyncConfig() SyncConfig {
+	return SyncConfig{
+		Batch:  randvar.NewFlag("5"),
+		Values: randvar.NewBytesFlag("uniform:60-80/1.0"),
+	}
+}
+
+// RunSync runs the sync benchmark.
+func RunSync(dir string, common *CommonConfig, cfg *SyncConfig) {
+	reg := newHistogramRegistry()
+	var bytes atomic.Uint64
+	var lastBytes uint64
+
+	opts := pebble.Sync
+	if common.DisableWAL {
+		opts = pebble.NoSync
+	}
+
+	batchDist := cfg.Batch
+
+	RunTest(dir, common, Test{
+		Init: func(d DB) {},
+
+		Run: func(d DB, wg *sync.WaitGroup) {
+			limiter := common.RateLimiter
+
+			for range common.Concurrency {
+				latency := reg.Register("ops")
+				wg.Go(func() {
+					rand := rand.New(rand.NewPCG(0, uint64(time.Now().UnixNano())))
+					var raw []byte
+					var buf []byte
+					var block []byte
+					for {
+						wait(limiter)
+
+						start := time.Now()
+						b := d.NewBatch()
+						var n uint64
+						count := int(batchDist.Uint64(rand))
+						for range count {
+							block = cfg.Values.Bytes(rand, block)
+
+							if cfg.WALOnly {
+								if err := b.LogData(block, nil); err != nil {
+									log.Fatal(err)
+								}
+							} else {
+								raw = encodeUint32Ascending(raw[:0], rand.Uint32())
+								key := cockroachkvs.EncodeMVCCKey(buf[:0], raw, 0, 0)
+								buf = key[:0]
+								if err := b.Set(key, block, nil); err != nil {
+									log.Fatal(err)
+								}
+							}
+							n += uint64(len(block))
+						}
+						if err := b.Commit(opts); err != nil {
+							log.Fatal(err)
+						}
+						latency.Record(time.Since(start))
+						bytes.Add(n)
+					}
+				})
+			}
+		},
+
+		Tick: func(elapsed time.Duration, i int) {
+			if i%20 == 0 {
+				fmt.Println("_elapsed____ops/sec___mb/sec__p50(ms)__p95(ms)__p99(ms)_pMax(ms)")
+			}
+			reg.Tick(func(tick histogramTick) {
+				h := tick.Hist
+				n := bytes.Load()
+				fmt.Printf("%8s %10.1f %8.1f %8.1f %8.1f %8.1f %8.1f\n",
+					time.Duration(elapsed.Seconds()+0.5)*time.Second,
+					float64(h.TotalCount())/tick.Elapsed.Seconds(),
+					float64(n-lastBytes)/(1024.0*1024.0)/tick.Elapsed.Seconds(),
+					time.Duration(h.ValueAtQuantile(50)).Seconds()*1000,
+					time.Duration(h.ValueAtQuantile(95)).Seconds()*1000,
+					time.Duration(h.ValueAtQuantile(99)).Seconds()*1000,
+					time.Duration(h.ValueAtQuantile(100)).Seconds()*1000,
+				)
+				lastBytes = n
+			})
+		},
+
+		Done: func(elapsed time.Duration) {
+			fmt.Println("\n_elapsed___ops(total)_ops/sec(cum)_mb/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)")
+			reg.Tick(func(tick histogramTick) {
+				h := tick.Cumulative
+				fmt.Printf("%7.1fs %12d %12.1f %11.1f %8.1f %8.1f %8.1f %8.1f %8.1f\n\n",
+					elapsed.Seconds(), h.TotalCount(),
+					float64(h.TotalCount())/elapsed.Seconds(),
+					float64(bytes.Load()/(1024.0*1024.0))/elapsed.Seconds(),
+					time.Duration(h.Mean()).Seconds()*1000,
+					time.Duration(h.ValueAtQuantile(50)).Seconds()*1000,
+					time.Duration(h.ValueAtQuantile(95)).Seconds()*1000,
+					time.Duration(h.ValueAtQuantile(99)).Seconds()*1000,
+					time.Duration(h.ValueAtQuantile(100)).Seconds()*1000)
+			})
+		},
+	})
+}
